+++ src/configuration.cpp
@@ -580,9 +580,9 @@ int art::register_valkey_configuration(ValkeyModuleCtx *ctx) {
                                              GetMaxMemoryRatio, SetMaxMemoryBytes, ApplyMaxMemoryRatio, nullptr);
     ret |= ValkeyModule_RegisterStringConfig(ctx, "min_fragmentation_ratio", "0.5", VALKEYMODULE_CONFIG_DEFAULT,
                                              GetMinFragmentation, SetMinFragmentation, ApplyMinFragmentation, nullptr);
-    ret |= ValkeyModule_RegisterStringConfig(ctx, "active_defrag", "off", VALKEYMODULE_CONFIG_DEFAULT,
+    ret |= ValkeyModule_RegisterStringConfig(ctx, "active_defrag", "on", VALKEYMODULE_CONFIG_DEFAULT,
                                              GetActiveDefragType, SetActiveDefragType, ApplyActiveDefragType, nullptr);
-    ret |= ValkeyModule_RegisterStringConfig(ctx, "maintenance_poll_delay", "10", VALKEYMODULE_CONFIG_DEFAULT,
+    ret |= ValkeyModule_RegisterStringConfig(ctx, "maintenance_poll_delay", "40", VALKEYMODULE_CONFIG_DEFAULT,
                                              GetMaintenancePollDelay, SetMaintenancePollDelay,
                                              ApplyMaintenancePollDelay, nullptr);
     ret |= ValkeyModule_RegisterStringConfig(ctx, "max_defrag_page_count", "10", VALKEYMODULE_CONFIG_DEFAULT,